HH-87 <update image_gallery_translation and image_gallery entity, repo, dto, mapper, service and controller>
